@@ -4,18 +4,18 @@
   "private": true,
   "dependencies": {
     "@testing-library/jest-dom": "^5.16.1",
-    "@testing-library/react": "^12.1.2",
+    "@testing-library/react": "^13.3.0",
     "@testing-library/user-event": "^13.5.0",
     "@types/jest": "^27.4.0",
     "@types/node": "^16.11.19",
     "@types/react": "^17.0.38",
-    "@types/react-dom": "^17.0.11",
-    "@types/react-redux": "^7.1.21",
+    "@types/react-dom": "^18.0.5",
     "@types/react-router": "^5.1.17",
     "@types/react-router-dom": "^5.3.2",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-redux": "^7.2.6",
+    "axios": "^0.25.0",
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0",
+    "react-query": "^3.39.1",
     "react-router": "^6.2.1",
     "react-router-dom": "^6.2.1",
     "react-scripts": "5.0.0",
@@ -27,6 +27,7 @@
     "build": "react-scripts build",
     "lint:test": "eslint src",
     "lint:fix": "eslint src --fix",
+    "format": "prettier --write \"./src/**/**.{ts,tsx,js,jsx,json}\"",
     "test": "react-scripts test",
     "eject": "react-scripts eject"
   },
@@ -51,11 +52,13 @@
   "devDependencies": {
     "@typescript-eslint/eslint-plugin": "^5.9.0",
     "@typescript-eslint/parser": "^5.9.0",
+    "autoprefixer": "^10.4.2",
     "eslint": "^8.6.0",
     "eslint-config-prettier": "^8.3.0",
     "eslint-plugin-prettier": "^4.0.0",
     "eslint-plugin-react": "^7.28.0",
+    "postcss": "^8.4.5",
     "prettier": "^2.5.1",
-    "tailwindcss": "^3.0.11"
+    "tailwindcss": "^3.0.12"
   }
 }
