@@ -1,6 +1,14 @@
 import React from 'react'
-import ReactDOM from 'react-dom'
+import ReactDOM from 'react-dom/client'
 import './index.css'
 import App from 'app'
 
-ReactDOM.render(<App />, document.getElementById('root'))
+const container = document.getElementById('root')
+
+const root = ReactDOM.createRoot(container!)
+
+root.render(
+  <React.StrictMode>
+    <App />
+  </React.StrictMode>,
+)
