@@ -1,1 +1,1 @@
-export {default as Spin} from './Spin'
+export {default as Spin} from './spin/Spin'
